+++ pydoclint/flake8_entry.py
@@ -249,12 +249,16 @@ class Plugin:
             ),
         )
         parser.add_option(
-            '-aicvd',
-            '--allow-inline-class-var-docs',
+            '-ricvd',
+            '--require-inline-class-var-docs',
             action='store',
             default='False',
             parse_from_config=True,
-            help='If True, allow inline documentation for class attributes.',
+            help=(
+                'If True, require inline documentation for class'
+                ' attributes. If False, require them to be documented'
+                ' in the class docstring instead.'
+            ),
         )
         parser.add_option(
             '-sdsa',
@@ -355,7 +359,9 @@ class Plugin:
         cls.only_attrs_with_ClassVar_are_treated_as_class_attrs = (
             options.only_attrs_with_ClassVar_are_treated_as_class_attrs
         )
-        cls.allow_inline_class_var_docs = options.allow_inline_class_var_docs
+        cls.require_inline_class_var_docs = (
+            options.require_inline_class_var_docs
+        )
         cls.should_document_star_arguments = (
             options.should_document_star_arguments
         )
@@ -449,9 +455,9 @@ class Plugin:
             '--only-attrs-with-ClassVar-are-treated-as-class-attrs',
             self.only_attrs_with_ClassVar_are_treated_as_class_attrs,
         )
-        allowInlineClassVarDocs = self._bool(
-            '--allow-inline-class-var-docs',
-            self.allow_inline_class_var_docs,
+        requireInlineClassVarDocs = self._bool(
+            '--require-inline-class-var-docs',
+            self.require_inline_class_var_docs,
         )
         shouldDocumentStarArguments = self._bool(
             '--should-document-star-arguments',
@@ -503,7 +509,7 @@ class Plugin:
             onlyAttrsWithClassVarAreTreatedAsClassAttrs=(
                 onlyAttrsWithClassVarAreTreatedAsClassAttrs
             ),
-            allowInlineClassVarDocs=allowInlineClassVarDocs,
+            requireInlineClassVarDocs=requireInlineClassVarDocs,
             shouldDocumentStarArguments=shouldDocumentStarArguments,
             omitStarsWhenDocumentingVarargs=omitStarsWhenDocumentingVarargs,
             checkStyleMismatch=checkStyleMismatch,