+++ pydoclint/main.py
@@ -290,12 +290,16 @@ def validateNativeModeNoqaLocation(
     ),
 )
 @click.option(
-    '-aicvd',
-    '--allow-inline-class-var-docs',
+    '-ricvd',
+    '--require-inline-class-var-docs',
     type=bool,
     show_default=True,
     default=False,
-    help='If True, allow inline documentation of class attributes.',
+    help=(
+        'If True, require inline documentation of class attributes.'
+        ' If False, require them to be documented in the class docstring'
+        ' instead.'
+    ),
 )
 @click.option(
     '-sdsa',
@@ -477,7 +481,7 @@ def main(  # noqa: C901, PLR0915
         require_return_section_when_returning_nothing: bool,
         require_yield_section_when_yielding_nothing: bool,
         only_attrs_with_classvar_are_treated_as_class_attrs: bool,
-        allow_inline_class_var_docs: bool,
+        require_inline_class_var_docs: bool,
         should_document_star_arguments: bool,
         omit_stars_when_documenting_varargs: bool,
         should_declare_assert_error_if_assert_statement_exists: bool,
@@ -589,7 +593,7 @@ def main(  # noqa: C901, PLR0915
         onlyAttrsWithClassVarAreTreatedAsClassAttrs=(
             only_attrs_with_classvar_are_treated_as_class_attrs
         ),
-        allowInlineClassVarDocs=allow_inline_class_var_docs,
+        requireInlineClassVarDocs=require_inline_class_var_docs,
         requireReturnSectionWhenReturningNothing=(
             require_return_section_when_returning_nothing
         ),
@@ -746,7 +750,7 @@ def _checkPaths(
         shouldDocumentPrivateClassAttributes: bool = False,
         treatPropertyMethodsAsClassAttributes: bool = False,
         onlyAttrsWithClassVarAreTreatedAsClassAttrs: bool = False,
-        allowInlineClassVarDocs: bool = False,
+        requireInlineClassVarDocs: bool = False,
         requireReturnSectionWhenReturningNothing: bool = False,
         requireYieldSectionWhenYieldingNothing: bool = False,
         shouldDocumentStarArguments: bool = True,
@@ -809,7 +813,7 @@ def _checkPaths(
             onlyAttrsWithClassVarAreTreatedAsClassAttrs=(
                 onlyAttrsWithClassVarAreTreatedAsClassAttrs
             ),
-            allowInlineClassVarDocs=allowInlineClassVarDocs,
+            requireInlineClassVarDocs=requireInlineClassVarDocs,
             requireReturnSectionWhenReturningNothing=(
                 requireReturnSectionWhenReturningNothing
             ),
@@ -848,7 +852,7 @@ def _checkFile(
         shouldDocumentPrivateClassAttributes: bool = False,
         treatPropertyMethodsAsClassAttributes: bool = False,
         onlyAttrsWithClassVarAreTreatedAsClassAttrs: bool = False,
-        allowInlineClassVarDocs: bool = False,
+        requireInlineClassVarDocs: bool = False,
         requireReturnSectionWhenReturningNothing: bool = False,
         requireYieldSectionWhenYieldingNothing: bool = False,
         shouldDocumentStarArguments: bool = True,
@@ -906,7 +910,7 @@ def _checkFile(
         onlyAttrsWithClassVarAreTreatedAsClassAttrs=(
             onlyAttrsWithClassVarAreTreatedAsClassAttrs
         ),
-        allowInlineClassVarDocs=allowInlineClassVarDocs,
+        requireInlineClassVarDocs=requireInlineClassVarDocs,
         requireReturnSectionWhenReturningNothing=(
             requireReturnSectionWhenReturningNothing
         ),